Update to packages and config
