+++ cfg/config.json
@@ -122,7 +122,7 @@
                "name": "muddy_floor",
                "type": "tile",
                "data": {
-                  "11561": 10000
+                  "11561": 20
                }
             }
          ]
@@ -139,7 +139,7 @@
                "name": "sand",
                "type": "tile",
                "data": {
-                  "231": 1
+                  "231": 10
                }
             },
             {
@@ -185,7 +185,7 @@
                "name": "dirt",
                "type": "tile",
                "data": {
-                  "103": 1
+                  "103": 25
                }
             }
          ]
@@ -243,7 +243,7 @@
                "name": "jungle_grass",
                "type": "tile",
                "data": {
-                  "3263": 1
+                  "3263": 20
                }
             },
             {
@@ -268,7 +268,7 @@
                "name": "jungle_grass",
                "type": "tile",
                "data": {
-                  "3263": 1
+                  "3263": 40
                }
             },
             {
@@ -375,7 +375,7 @@
                "name": "dirt",
                "type": "tile",
                "data": {
-                  "103": 1
+                  "103": 20
                }
             },
             {
@@ -383,11 +383,11 @@
                "type": "tile",
                "data": {
                   "20776": 2000,
-                  "20777": 1000,
-                  "20778": 1200,
-                  "20779": 1200,
-                  "20780": 2500,
-                  "20781": 2000
+                  "20777": 100,
+                  "20778": 120,
+                  "20779": 120,
+                  "20780": 250,
+                  "20781": 200
                }
             }
          ]
@@ -426,19 +426,19 @@
                "name": "dirt",
                "type": "tile",
                "data": {
-                  "103": 1
+                  "103": 30
                }
             },
             {
                "name": "forest_floor",
                "type": "tile",
                "data": {
-                  "20776": 2000,
-                  "20777": 1000,
-                  "20778": 1200,
-                  "20779": 1200,
-                  "20780": 2500,
-                  "20781": 2000
+                  "20776": 1800,
+                  "20777": 20,
+                  "20778": 25,
+                  "20779": 20,
+                  "20780": 25,
+                  "20781": 20
                }
             }
          ]
@@ -477,7 +477,7 @@
                "name": "dirt",
                "type": "tile",
                "data": {
-                  "103": 1
+                  "103": 25
                }
             },
             {
